feat!: rename generator from servus:event_handler to servus:event
- Generator now creates app/events/name.rb (no _handler suffix)
- Spec placed at spec/app/events/name_spec.rb
- Generated class inherits from ApplicationEvent
- Event name inferred from class name (no explicit event_name call)
- Template uses new Event DSL and emit_event matcher
